Move ExtensionContext.environmentVariableCollection to stable
Fixes #46696
